pull out user struct into model package as example
