+++ internal/models/user.go
@@ -1,0 +1,19 @@
+package models
+
+type User struct {
+	Username      string `json:"username"`
+	ID            int    `json:"id"`
+	Email         string `json:"email"`
+	FirstName     string `json:"first_name"`
+	LastName      string `json:"last_name"`
+	AccountNumber string `json:"account_number"`
+	AddressString string `json:"address_string"`
+	IsActive      bool   `json:"is_active"`
+	IsOrgAdmin    bool   `json:"is_org_admin"`
+	IsInternal    bool   `json:"is_internal"`
+	Locale        string `json:"locale"`
+	OrgID         string `json:"org_id"`
+	DisplayName   string `json:"display_name"`
+	Type          string `json:"type"`
+	Entitlements  string `json:"entitlements"`
+}
