+++ internal/service/catchall/ephemeral.go
@@ -13,28 +13,12 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/redhatinsights/mbop/internal/models"
 	"golang.org/x/oauth2/clientcredentials"
 )
 
 // TODO: move these to the models package (internal/models/) if we can reuse
 // them, otherwise label them appropriately.
-type User struct {
-	Username      string `json:"username"`
-	ID            int    `json:"id"`
-	Email         string `json:"email"`
-	FirstName     string `json:"first_name"`
-	LastName      string `json:"last_name"`
-	AccountNumber string `json:"account_number"`
-	AddressString string `json:"address_string"`
-	IsActive      bool   `json:"is_active"`
-	IsOrgAdmin    bool   `json:"is_org_admin"`
-	IsInternal    bool   `json:"is_internal"`
-	Locale        string `json:"locale"`
-	OrgID         string `json:"org_id"`
-	DisplayName   string `json:"display_name"`
-	Type          string `json:"type"`
-	Entitlements  string `json:"entitlements"`
-}
 
 type JSONStruct struct {
 	PublicKey       string `json:"public_key"`
@@ -50,13 +34,13 @@ type usersByInput struct {
 }
 
 type Resp struct {
-	User      User   `json:"user"`
-	Mechanism string `json:"mechanism"`
+	User      models.User `json:"user"`
+	Mechanism string      `json:"mechanism"`
 }
 
 type AccV2Resp struct {
-	Users     []User `json:"users"`
-	UserCount int    `json:"userCount"`
+	Users     []models.User `json:"users"`
+	UserCount int           `json:"userCount"`
 }
 
 type Realm struct {
@@ -68,7 +52,7 @@ type V1UserInput struct {
 	Users []string `json:"users"`
 }
 
-func (m *MBOPServer) findUserByID(username string) (*User, error) {
+func (m *MBOPServer) findUserByID(username string) (*models.User, error) {
 	users, err := m.getUsers()
 
 	if err != nil {
@@ -83,14 +67,14 @@ func (m *MBOPServer) findUserByID(username string) (*User, error) {
 	return nil, fmt.Errorf("User is not known")
 }
 
-func (m *MBOPServer) findUsersBy(accountNo string, orgID string, adminOnly string, status string, limit int, sortOrder string, queryBy string, input *usersByInput, users *V1UserInput) ([]User, error) {
+func (m *MBOPServer) findUsersBy(accountNo string, orgID string, adminOnly string, status string, limit int, sortOrder string, queryBy string, input *usersByInput, users *V1UserInput) ([]models.User, error) {
 	usersList, err := m.getUsers()
 
 	if err != nil {
 		return nil, err
 	}
 
-	out := []User{}
+	out := []models.User{}
 	for _, user := range usersList {
 		// When adminOnly is true, parameter “status” is ignored
 		if adminOnly == "true" && !user.IsOrgAdmin {
@@ -186,19 +170,19 @@ func (m *MBOPServer) getJWT(realm string) (*JSONStruct, error) {
 	return jsonstruct, nil
 }
 
-func (m *MBOPServer) getUser(w http.ResponseWriter, r *http.Request) (*User, error) {
+func (m *MBOPServer) getUser(w http.ResponseWriter, r *http.Request) (*models.User, error) {
 	auth := r.Header.Get("Authorization")
 	if auth == "" {
-		return &User{}, fmt.Errorf("no auth header found")
+		return &models.User{}, fmt.Errorf("no auth header found")
 	}
 	if !strings.Contains(auth, "Basic") {
-		return &User{}, fmt.Errorf("auth header is not basic")
+		return &models.User{}, fmt.Errorf("auth header is not basic")
 	}
 
 	data, err := base64.StdEncoding.DecodeString(auth[6:])
 
 	if err != nil {
-		return &User{}, fmt.Errorf("could not split header")
+		return &models.User{}, fmt.Errorf("could not split header")
 	}
 	parts := strings.Split(string(data), ":")
 
@@ -206,7 +190,7 @@ func (m *MBOPServer) getUser(w http.ResponseWriter, r *http.Request) (*User, err
 	password := parts[1]
 
 	if err != nil {
-		return &User{}, fmt.Errorf("can't create keycloak client: %s", err.Error())
+		return &models.User{}, fmt.Errorf("can't create keycloak client: %s", err.Error())
 	}
 
 	oauthClientConfig := clientcredentials.Config{
@@ -220,19 +204,19 @@ func (m *MBOPServer) getUser(w http.ResponseWriter, r *http.Request) (*User, err
 	resp, err := k.Get(m.getURL("/auth/realms/redhat-external/account/"))
 
 	if err != nil {
-		return &User{}, fmt.Errorf("couldn't auth user: %s", err.Error())
+		return &models.User{}, fmt.Errorf("couldn't auth user: %s", err.Error())
 	}
 
 	defer resp.Body.Close()
 
 	if resp.StatusCode != 200 {
-		return &User{}, fmt.Errorf("user unauthorized: %d", resp.StatusCode)
+		return &models.User{}, fmt.Errorf("user unauthorized: %d", resp.StatusCode)
 	}
 
 	userObj, err := m.findUserByID(username)
 
 	if err != nil {
-		return &User{}, fmt.Errorf("couldn't find user: %s", err.Error())
+		return &models.User{}, fmt.Errorf("couldn't find user: %s", err.Error())
 	}
 	return userObj, nil
 }
@@ -310,7 +294,7 @@ type usersSpec struct {
 	Attributes map[string][]string `json:"attributes"`
 }
 
-func (m *MBOPServer) getUsers() (users []User, err error) {
+func (m *MBOPServer) getUsers() (users []models.User, err error) {
 	resp, err := m.Client.Get(m.getURL("/auth/admin/realms/redhat-external/users", map[string]string{"max": "2000"}))
 	if err != nil {
 		fmt.Printf("\n\n%s\n\n", err.Error())
@@ -331,7 +315,7 @@ func (m *MBOPServer) getUsers() (users []User, err error) {
 		return nil, err
 	}
 
-	users = []User{}
+	users = []models.User{}
 
 	for _, user := range *obj {
 		IsActiveRaw := user.Attributes["is_active"][0]
@@ -357,7 +341,7 @@ func (m *MBOPServer) getUsers() (users []User, err error) {
 			entitle = user.Attributes["entitlements"][0]
 		}
 
-		users = append(users, User{
+		users = append(users, models.User{
 			Username:      user.Username,
 			ID:            ID,
 			Email:         user.Email,
